move db table count worker to metrics folder
